Continued abit of stuff in the add lesson where now i added how to add students to a lesson
I think students should not have a field for lessons and that it should be seperate
instead we can have maybe a record associated with each student for each lesson they attend to make the financial ccounting easier(havent done this yet)
the plus sign on the choose student page is for quick add of students. I want to add another seperate page just for viewing editing deleting and creating new students
using the drawer we pull from the left idea. also for the list of students i enveloped them in cards. If you think they look nice i can do the same for the lessons as well.
the code is a bit more complicated because i used a recyclerview which is like a list view and is supposed to be dynamically updated and so we wont have to use update list

similiarly i will be using this recyclerview for the lessons as well so they will be dynamically updated after we just add a lesson