Ok add lesson should definitely work now. Only two minor aspects not implemented yet:
If we want to add multiple students we will have to currently add one by one. Each time the student clicked will be added. Right now there is no checking of whether the student clicked is already selected so I will implement that checking later.
Also will make it such that the currently selected students appear at the top, but this will be harder to do. But like that it will look much better.
Also plan to make it such that if you long press you can change to multiple select mode but I think that too is not that high priority as we currently can add multiple one by one. Also notice even though we select multiple times only the first student
selected appears. I will add a +1, +2 etc to show how many more students are there in addition so its visually clearer.

The other aspect is that if client chooses to change the selected date when adding new lesson. I will need to change the selected daymap and or monthmap if need to but for now lets just assume that even though the option is  there the client doesnt change the selected date
within the add lesson interface.

i will make the view lesson interface work later by using the recycler view again later