Changed the look of the calendar page to white so that we wont have the orange background being darker and lighter and so it will be seamless with the background of the calendar itself.
Modified the view lessons to use a recycler view to make it look abit better. Also tried to work on the sms manager will need to test that. I added a field under lessons that is a boolean that
is true if checked in. Is it possible to stop the notifications from appearing once this boolean is set to true? Once that is done can populate the lessons that are awaiting a summary report