Created a RecurringLesson and RecurringLessonMap to handle recurring lessons (not using equals).
Notifications will now update with number of lessons in queue on the drawer. Can change to other text. Expired lessons will automatically be deleted and caldroid will be refreshed on delete.

Fixed a bug made by myself for lessons using 12h instead of 24h (hour instead of hour of day) in lesson.